backend: Added db to request ctx, rather than custom
This way of attaching the Koa ctx object to the
server request context allows for extending the capabilities of
resolvers.

TODO: time for Passport

+++ backend/lib/resolvers/Mutation.js
@@ -5,11 +5,11 @@
  * @param {*} args => {
  *      name: String
  * }, only the name should be accept, everything else vanilla
- * @param {*} ctx 
+ * @param {*} { dao }
  * @param {*} info 
  */
-const createSpace = async (_, args, ctx, info) => {
-    const { SpacesDAO } = ctx.DAO; //grabs DAO from db ctx
+const createSpace = async (_, args, { dao }, info) => {
+    const { SpacesDAO } = dao; //grabs DAO from db ctx
     try{
         const space = await SpacesDAO.create(args);
         return space;
@@ -22,8 +22,15 @@ const createSpace = async (_, args, ctx, info) => {
 
 }
 
-const updateSpace = async (_, args, ctx, info) => {
-    const { SpacesDAO } = ctx.DAO;
+/**
+ * 
+ * @param {*} _ 
+ * @param {*} args 
+ * @param {*} param2 
+ * @param {*} info 
+ */
+const updateSpace = async (_, args, { dao }, info) => {
+    const { SpacesDAO } = dao;
     try {
         const space = SpacesDAO.update(args);
         return space;
@@ -32,13 +39,23 @@ const updateSpace = async (_, args, ctx, info) => {
     }
 }
 
-const deleteSpace = async (_, args, ctx, info) => {
-    const { SpacesDAO, CardsDAO } = ctx.DAO;
+/**
+ * 
+ * @param {*} _ 
+ * @param {*} args 
+ * @param {*} param2 
+ * @param {*} info 
+ */
+const deleteSpace = async (_, args, { dao }, info) => {
+    const { SpacesDAO, CardsDAO } = dao;
     try {
         const space = await SpacesDAO.delete(args);
-        space.cards.forEach(cardId => {
-            CardsDAO.delete({ id: cardId })
-        })
+        // Check to see if cards are not null
+        if(space.cards){
+            space.cards.forEach(cardId => {
+                CardsDAO.delete({ id: cardId })
+            })
+        }
         return space;
     } catch (err) {
         console.log(err);
@@ -60,12 +77,12 @@ const deleteSpace = async (_, args, ctx, info) => {
  *      } // Object w/ diff values
  *      
  * }
- * @param {*} ctx => holds db ctx
+ * @param {*}{ dao } => holds db ctx
  * @param {*} info 
  */
-const createCard = async (_, args, ctx, info) => {
-    //const { Spaces, Cards } = ctx.DAO; //grabs model from db ctx
-    const { SpacesDAO, CardsDAO } = ctx.DAO;
+const createCard = async (_, args, { dao }, info) => {
+    //const { Spaces, Cards } = dao; //grabs model from db ctx
+    const { SpacesDAO, CardsDAO } = dao;
     //TODO: Write validation
     //Assume user is perfect... for now...
     try {
@@ -83,8 +100,16 @@ const createCard = async (_, args, ctx, info) => {
     }
 }
 
-const updateCard = async (_, args, ctx, info) => {
-    const { CardsDAO } = ctx.DAO;
+
+/**
+ * 
+ * @param {*} _ 
+ * @param {*} args 
+ * @param {*} param2 
+ * @param {*} info 
+ */
+const updateCard = async (_, args, { dao }, info) => {
+    const { CardsDAO } = dao;
     // TODO: Data validation
     try {
         const card = await CardsDAO.update(args);
@@ -101,11 +126,11 @@ const updateCard = async (_, args, ctx, info) => {
  * @param {*} args => {
  *      id, // Card belongs to a Space (this is a space ID)
  * }
- * @param {*} ctx => holds db ctx
+ * @param {*} { dao } => holds db ctx
  * @param {*} info 
  */
-const deleteCard = async (_, args, ctx, info) => {
-    const { SpacesDAO, CardsDAO } = ctx.DAO; //grabs model from db ctx
+const deleteCard = async (_, args, { dao }, info) => {
+    const { SpacesDAO, CardsDAO } = dao; //grabs model from db ctx
     //TODO: Write validation
     //Assume user is perfect... for now...
     try {
@@ -121,13 +146,12 @@ const deleteCard = async (_, args, ctx, info) => {
     }
 }
 
+
 module.exports = {
-    Mutation: {
-        createSpace,
-        updateSpace,
-        deleteSpace,
-        createCard,
-        updateCard,
-        deleteCard
-    }
+    createSpace,
+    updateSpace,
+    deleteSpace,
+    createCard,
+    updateCard,
+    deleteCard
 }
